fix: add custom hook for wknd app url
