+++ components/Header.jsx
@@ -1,7 +1,5 @@
 import Image from "next/image";
-import { useRouter } from "next/router";
-import { useMemo } from "react";
-import { getWkndAppUrl } from "./utils";
+import { useWkndAppUrl } from "../hooks";
 
 const ArrowSvg = () => (
   <svg xmlns="http://www.w3.org/2000/svg" height="18" viewBox="0 0 18 18" width="18">
@@ -15,10 +13,7 @@ const ArrowSvg = () => (
 );
 
 export default function Header({ isAuthorVersion, host }) {
-  const router = useRouter();
-  const queryParams = router.query;
-  const wkndAppUrl = useMemo(() => getWkndAppUrl(queryParams), [queryParams]);
-
+  const wkndAppUrl = useWkndAppUrl();
   return (
     <header className="header">
       <div className="content">
